Minor tweaks to llama.cpp command

+++ start-server.sh
@@ -11,13 +11,20 @@ docker run -d --gpus all \
   ghcr.io/ggml-org/llama.cpp:server-cuda \
   -m /models/gemma-4-26B-A4B-it-UD-Q4_K_M.gguf \
   --slot-save-path /cache \
-  --ctx-size 131072 --no-mmap \
-  --host 0.0.0.0 --port 8080 --n-gpu-layers 999 \
+  --ctx-size 65536 --no-mmap --mlock \
+  --host 0.0.0.0 --port 8080 --n-gpu-layers 99 \
   --flash-attn on \
-  -np 1
+  --threads 2 --alias gemma4 \
+  -np 1 --cache-type-k q4_0 --cache-type-v q4_0 --keep -1
 
 echo "Waiting for llama.cpp to load model"
 
+sleep 10
+# If we 404 here, abort with error
+if [ $(curl -s -o /dev/null -w '%{http_code}' "http://localhost:$PORT/health") -eq 404 ]; then
+    echo "Error: Health check returned 404. Aborting."
+    exit 1
+fi
 # Loop until the health endpoint returns a 200 OK
 until [ $(curl -s -o /dev/null -w '%{http_code}' "http://localhost:$PORT/health") -eq 200 ]; do
     printf '.'
